requirements.txt
argon2-cffi==19.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
cffi==1.12.3
Django==2.2.4
django-bootstrap4==1.0.1
django-braces==1.13.0
django-crispy-forms==1.7.2
Faker==2.0.1
gunicorn==19.9.0
Pillow==6.1.0
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
text-unidecode==1.2